% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative.r
\name{alignment_region}
\alias{alignment_region}
\title{Retrieve a genomic alignment from a given region}
\usage{
alignment_region(region, species = "homo_sapiens", aligned = FALSE,
  compact = TRUE, compara = "multi", display_species_set = NULL,
  mask = "soft", method = "EPO", species_set = NULL,
  species_set_group = "mammals", req_format = "json")
}
\arguments{
\item{region}{given as chr:start-end}
}
\value{
tree and information about sequence of the region for each species
}
\description{
Retrieve a genomic alignment from a given region
}

