% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caroline.r
\name{dnds}
\alias{dnds}
\title{Get dN/dS data for a given gene symbol}
\usage{
dnds(symbol)
}
\arguments{
\item{symbol}{gene symbol}
}
\value{
a dataframe contaning dnds, species and taxonomic level
}
\description{
Get dN/dS data for a given gene symbol
}

