% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparative.r
\name{genetree_member_symbol}
\alias{genetree_member_symbol}
\title{Retreive a gene tree from species and gene symbol}
\usage{
genetree_member_symbol(species, symbol, tree_format, aligned = NULL,
  external_db = NULL, nh_format = NULL, object_type = NULL,
  sequence = NULL)
}
\description{
Retreive a gene tree from species and gene symbol
}

