% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.r
\name{lookup_symbol}
\alias{lookup_symbol}
\title{Retrieve information about a sequence via gene symbol and species}
\usage{
lookup_symbol(symbol, expand = FALSE, species = "homo_sapiens",
  format = "full", return_format = "json")
}
\arguments{
\item{symbol}{character, gene symbol(s) from which to retreive information.}

\item{species}{character, species from which to retreieve information. See
\code{info_species} for a list of avaliable species. Defaults to homo_sapiens}

\item{format, }{character, type of record to return either 'full' or 'condensed'}

\item{return_format, }{character method by which record is returned. Defaults
to \code{json}}

\item{symbol}{logical, if \code{TRUE} return information about transcripts/proteins
associated with these record. Defaults to \code{FALSE}}
}
\description{
Retrieve information about a sequence via gene symbol and species
}
\examples{
human_brca <- lookup_symbol("BRCA2")
chimp_brca <- lookup_symbol("BRCA2", species="ptro")
chimp_brca_expanded <- lookup_symbol("BRCA2", species="ptro", expand=TRUE)
}
\references{
\url{lhttp://rest.ensembl.org/documentation/info/symbol_lookup}
}

