% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapping.r
\name{map_genomic_coordinates}
\alias{map_genomic_coordinates}
\title{Convert  coordinates between gene}
\usage{
map_genomic_coordinates(id, type, region, format = "json")
}
\description{
Convert  coordinates between gene
}

