% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vep.r
\name{vep_allele}
\alias{vep_allele}
\title{Fetch variant consequences fro a given allele at a given site}
\usage{
vep_allele(species = "hsap", region, allele, format = "json")
}
\description{
Fetch variant consequences fro a given allele at a given site
}
\examples{
vep_allele(species="hsap", region="9:22125503-22125502:1", allele="C")
}

