% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vep.r
\name{vep_id}
\alias{vep_id}
\title{Get all variants and effects}
\usage{
vep_id(id, species = "hsap", format = "json")
}
\arguments{
\item{id}{gene ID}

\item{species}{ensembl name for the species from which the gene ID comes
(default = hsap = humans)}

\item{format}{format for fetched recrod (default = json)}
}
\description{
Get all variants and effects
}
\examples{
vep_id('COSM467')
vep_id(c("rs116035550", "COSM476"))
}

