% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep.r
\name{vep_region}
\alias{vep_region}
\title{Fetch variant consequences from regions}
\usage{
vep_region(region, species = "human", format = "json")
}
\description{
Fetch variant consequences from regions
}
\examples{
vep_region(c("21 26960070 rs116645811 G A", "21 26965148 - G A"))
}

