% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparative.r
\name{alignment_region}
\alias{alignment_region}
\title{Retreieve a genomic alignment from a given region}
\usage{
alignment_region(region, species = "homo_sapians", aligned = FALSE,
  compact = TRUE, compara = "multi", display_species_set = NULL,
  mask = "soft", method = "EPO", species_set = NULL,
  species_set_group = "mammals", req_format = "json")
}
\description{
Retreieve a genomic alignment from a given region
}

